--- a/hawkular-inventory-api/src/main/java/org/hawkular/inventory/api/Data.java
+++ b/hawkular-inventory-api/src/main/java/org/hawkular/inventory/api/Data.java
@@ -139,8 +139,8 @@ public final class Data {
 
         @Override
         default boolean anyExists() {
-            return !flatData(RelativePath.empty().get(),
-                    Pager.builder().withPageSize(1).orderBy(Order.unspecified()).build()).isEmpty();
+            return flatData(RelativePath.empty().get(), Pager.builder().withPageSize(1).orderBy(Order.unspecified())
+                    .build()).hasNext();
         }
     }
 }