--- a/hawkular-inventory-api/src/main/java/org/hawkular/inventory/api/ResolvableToMany.java
+++ b/hawkular-inventory-api/src/main/java/org/hawkular/inventory/api/ResolvableToMany.java
@@ -56,6 +56,6 @@ public interface ResolvableToMany<Entity> {
      * @return true if there is at least 1 entity on the current position in the inventory traversal
      */
     default boolean anyExists() {
-        return entities(Pager.builder().withPageSize(1).orderBy(Order.unspecified()).build()).hasNext();
+        return entities(Pager.single()).hasNext();
     }
 }