--- a/hawkular-inventory-api/src/main/java/org/hawkular/inventory/api/filters/RelationWith.java
+++ b/hawkular-inventory-api/src/main/java/org/hawkular/inventory/api/filters/RelationWith.java
@@ -64,7 +64,7 @@ public final class RelationWith {
         if (values == null || values.length == 0) {
             throw new IllegalArgumentException("there must be at least one value of the relation name");
         }
-        PropertyValues names = new PropertyValues("label", values);
+        PropertyValues names = new PropertyValues("label", (Object[]) values);
         return names;
     }