This PR is trying to speed up the REST api. 
by:
* Making the page based on Iterable instead of List
* Using Piped Streams + lazy JSON serialization when iterating through the result page

jProfiler output for our load tests:
![screenshot from 2015-09-07 23-49-47](https://cloud.githubusercontent.com/assets/535866/9731254/1fe24f64-561c-11e5-9895-b1f7b95334d7.png)

jProfiler output when security is ON:
![screenshot from 2015-09-05 01-33-36](https://cloud.githubusercontent.com/assets/535866/9731276/36eb7d2a-561c-11e5-8517-fa56a2d1c9f2.png)


before 

---- Global Information --------------------------------------------------------
 request count                                        476 (OK=476    KO=0     )
 min response time                                     51 (OK=51     KO=-     )
 max response time                                    362 (OK=362    KO=-     )
 **mean response time                                   187 (OK=187    KO=-     )**
 std deviation                                         42 (OK=42     KO=-     )
 response time 50th percentile                        187 (OK=187    KO=-     )
 response time 75th percentile                        214 (OK=214    KO=-     )
 mean requests/sec                                 12.689 (OK=12.689 KO=-     )
---- Response Time Distribution ------------------------------------------------
 t < 800 ms                                           476 (100%)
 800 ms < t < 1200 ms                                   0 (  0%)
 t > 1200 ms                                            0 (  0%)
 failed                                                 0 (  0%)

after

---- Global Information --------------------------------------------------------
 request count                                        476 (OK=461    KO=15    )
 min response time                                      5 (OK=9      KO=5     )
 max response time                                   1102 (OK=1102   KO=50    )
 **mean response time                                    85 (OK=87     KO=15    )**
 std deviation                                         61 (OK=60     KO=11    )
 response time 50th percentile                         84 (OK=86     KO=11    )
 response time 75th percentile                        112 (OK=113    KO=19    )
 mean requests/sec                                 21.664 (OK=20.981 KO=0.683 )
---- Response Time Distribution ------------------------------------------------
 t < 800 ms                                           460 ( 97%)
 800 ms < t < 1200 ms                                   1 (  0%)
 t > 1200 ms                                            0 (  0%)
 failed                                                15 (  3%)