@Jiri-Kremser 2015-09-08
> adding "don't merge yet" label, just so that forgetting to switch on the authz becomes ever so slightly less probable ;)

I was pretty sure I've added the WIP flag yesterday, but obviously not