@metlos 2015-09-10
btw, I'm not sure why the commit https://github.com/hawkular/hawkular-inventory/commit/fd77c3d81f8fa716e484888de72f3edba139a06e is still claimed to be part of this PR when it's already in master.