@Jiri-Kremser 2015-09-11
Btw. if I run the profiler against this PR the bottleneck is now moved into the `DrainedRangeFilterPipe.processNextStart()` because it's being called with each (or every second ["kazdy druhy"]) `hasNext()` and/or `next()`. But this particular pipe processes the whole content in each `processNextStart`.

 I think we are misusing it here. As far as I understand it well 1 call of `DrainedRangeFilterPipe.processNextStart()` is enough for obtaining the total count, but if the result is wrapped in this pipe and consumer iterates through it not efficient. Perhaps overriding the `hasNext()` and `next()` from `AbstractPipe` could do the trick, somehow..