@Jiri-Kremser 2015-09-18
ok I think I am done.
hawkular with the master version of inventory spends 497 seconds on discovery
with this change it's 152 seconds on my machine
(both with security, bus and everything)
-> 3.26 speedup